+++ packages/pubsub-gossipsub/src/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './shuffle.js'
-export * from './messageIdToString.js'
-export { getPublishConfigFromPeerId } from './publishConfig.js'